requirements.txt
fastapi
uvicorn
firebase-admin
pyrebase4
gunicorn
